refactor: 도메인들을 domain 패키지로 분리
