+++ src/main/kotlin/racingcar/Main.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.domain.Cars
+import racingcar.domain.RacingGame
 import racingcar.view.InputView
 import racingcar.view.ResultView
