+++ src/main/kotlin/racingcar/domain/PositionHistory.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 data class PositionHistory(
     val name: String,
