+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,11 +1,12 @@
-package racingcar
+package racingcar.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.Car.Companion.DEFAULT_POSITION
+import racingcar.domain.Car
+import racingcar.domain.Car.Companion.DEFAULT_POSITION
 import racingcar.strategy.ManualStrategy
 
 class CarTest {
