+++ src/test/kotlin/racingcar/domain/RoundHistoryTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
